Fix type of FLYWHEEL_ON_VELOCITY

+++ include/constants.hpp
@@ -5,7 +5,7 @@
 // namespace for nonphysical and programmatical constants
 namespace constants {
 constexpr float FLYWHEEL_PROPORTION = 1.0;  // (needed disc velocity) * (FLYWHEEL_PROPORTION) = (needed motor velocity)
-constexpr int16_t FLYWHEEL_ON_VELOCITY = 100;  // the default flywheel velocity when not aimed
+constexpr int32_t FLYWHEEL_ON_VELOCITY = 100;  // the default flywheel velocity when not aimed
 constexpr float GRAVITY = 9.79909f;  // local gravity in m/s/s
 constexpr float SENSOR_VERTICAL_FOV = 0.715585f;  // 41 deg to radians
 constexpr uint32_t PISTON_DELAY_TIME = 500;  // number of milliseconds to wait between extension and retraction of piston indexer
